Fix signed compare warnings
warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

conversions

Update src/ToolBox/superpmi/mcs/verbdumptoc.cpp



Commit migrated from https://github.com/dotnet/coreclr/commit/5a97d8e9f14763a4ff90d14d045f0173e8d4b944

+++ src/clr/src/ToolBox/SOS/Strike/util.cpp
@@ -1726,7 +1726,7 @@ int GetValueFieldOffset(CLRDATA_ADDRESS cdaMT, __in_z LPCWSTR wszFieldName, Dacp
     if (dmtd.ParentMethodTable)
     {
         DWORD retVal = GetValueFieldOffset(dmtd.ParentMethodTable, wszFieldName, pDacpFieldDescData);
-        if (retVal != NOT_FOUND)
+        if (retVal != (DWORD)NOT_FOUND)
         {
             // Return in case of error or success. Fall through for field-not-found.
             return retVal;
